requirements.txt
kornia==0.5.2
streamlit==0.79.0
matplotlib==3.3.2
torch==1.8.1
opencv_python_headless==4.5.1.48
numpy==1.19.5
torchvision==0.9.1
Pillow==8.2.0
